# Conflicts:
#	feature/maps/src/main/java/com/teamsolply/solply/maps/component/bottomsheet/AddCourseBottomSheet.kt
#	feature/onboarding/src/main/java/com/teamsolply/solply/onboarding/screen/SelectTownScreen.kt

+++ feature/place/src/main/java/com/teamsolply/solply/place/model/PlaceData.kt
@@ -3,7 +3,7 @@ package com.teamsolply.solply.place.model
 import com.teamsolply.solply.model.PlaceType
 
 data class PlaceData(
-    val placeId: Int,
+    val placeId: Long,
     val placeName: String,
     val thumbnailUrl: Int,
     val primaryTag: PlaceType,
